% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_Miracle.R
\name{Calculate_Miracle}
\alias{Calculate_Miracle}
\title{Mediators of Immune Response Against Cancer in soLid microEnvironments (MIRACLE)}
\usage{
Calculate_Miracle(matrix, platform = "ens", center = FALSE)
}
\arguments{
\item{matrix}{Numeric matrix of normalized gene expression values.
This matrix holds the "Gene/Transcript Features" on the rows and the "Samples" on the columns.
Normalization is typically carried out by methods like "EDASeq" for RNAseq data and "RMA" or "Quantile" methods for Microarrays.}

\item{platform}{Character string that shows the platform for the expression matrix. Can take values of c("ens", "u133p2", "entrez", "gene"). 
The default is "ens" for Ensembl Gene Ids (ie. ENSG000..). 
If the expression matrix is based on Affymetrix U133p2 or U133A platforms, please use "u133p2"
If the expression matrix is based on Entrez Gene Ids or Gene Symbols, please use "entrez" or "gene", respectively.
If the expression matrix is based on IlluminaHT12v3 or IlluminaHT12v4 platforms, please use "ilmn"
When using platforms other than "ens", there will be some loss of information due to features not matching or multi-matching.}

\item{center}{Logical of length 1. User can choose whether to mean center the expression matrix. Defaults to FALSE.}
}
\value{
data.frame of ICR (measure of immune infiltrate) and Miracle scores (Columns) for each sample (Rows).
}
\description{
Calculates Geneset Enrichment for ICR geneset and Miracle scores
}
\examples{
miracle_scores<-Calculate_Miracle(matrix)
}
\references{
\url{https://doi.org/10.1186/s40425-018-0355-5}
}
\author{
Tolga Turan, \email{tolga.turan@abbvie.com}
}
