% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geneset_Enrich.R
\name{Geneset_Enrich}
\alias{Geneset_Enrich}
\title{Geneset Enrichment (GSE)
Calculates single sample GSE scores using "yaGST" method}
\usage{
Geneset_Enrich(matrix1, geneset_list, verbose = TRUE, center = TRUE,
  n_cores = detectCores() - 2)
}
\arguments{
\item{geneset_list}{a List of gene/Transcript features. 
Preferably all or at least 1 feature in each geneset must be present in the rownames of the expression matrix.}

\item{verbose}{Logical of length 1. User can choose to display messages. Defaults to TRUE. 
If there are large numbers of genesets it can be set to FALSE}

\item{center}{Logical of length 1. User can choose whether to mean center the expression matrix. Defaults to TRUE.}

\item{n_cores}{Integer of length 1. Specifies the number of parallel cores running the function.
Defaults to detectCores()-2}

\item{matrix}{Numeric matrix of normalized gene expression values with unique column and rownames. 
    This matrix holds the "Gene/Transcript Features" on the rows and the "Samples" on the columns. 
Normalization is typically carried out by methods like "EDASeq" for RNAseq data and "RMA" or "Quantile" methods for Microarrays.}
}
\value{
data.frame holding geneset enrichment values for each geneset (columns) and for each sample (columns). 
If the geneset is of length 1, then a named vector of enrichment values.
}
\description{
Geneset Enrichment (GSE)
Calculates single sample GSE scores using "yaGST" method
}
\examples{
normalized_enrichment_scores<-Geneset_Enrich(matrix1, geneset_list1)
}
\references{
\url{http://github.com/miccec/yaGST}
}
\author{
Tolga Turan, \email{tolga.turan@abbvie.com}
}
