% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miracle.R
\name{miracle}
\alias{miracle}
\title{miracle is the main function}
\usage{
miracle(seqScore, mirtarList, mirExpr, tarExpr, exprFilter = 1,
  OutputSelect = TRUE, ID_pos = c())
}
\arguments{
\item{seqScore}{}

\item{mirtarList}{}

\item{mirExpr}{}

\item{tarExpr}{}

\item{exprFilter}{}

\item{OutputSelect}{}

\item{ID_pos}{}
}
\description{
miracle is the main function
}
