% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miracle.R
\name{miracle_ind}
\alias{miracle_ind}
\title{Title miRACLe predicts miRNA-mRNA interactions for individual samples as well
 as sample population}
\usage{
miracle_ind(seqScore, mirExpr, tarExpr, OutputSelect = TRUE)
}
\arguments{
\item{seqScore}{A Z x 3 data.frame that contains sequence-based interaction
scores for putative miRNA-mRNA pairs. The first column must contain gene
symbols for mRNAs, the second column must contain miRNA ids for miRNAs,
the third column must contain sequence-based score for the miRNA-mRNA pair.}

\item{mirExpr}{A numeric vector as the observed expression of M miRNAs. NOTE:
rownames is required for mirExpr_ind to map each miRNA to the miRNA id in
the second column of seqScore.}

\item{tarExpr}{A numeric vector as the observed expression of N mRNAs. NOTE:
rownames is required for tarExpr_ind to map each mRNA to the gene symbol in
the first column of seqScore.}

\item{OutputSelect}{Logical, select “TRUE” to return the top 10 percent
ranked predictions by scores, and “FALSE” to return the whole prediction
result. Default is TRUE}
}
\description{
Title miRACLe predicts miRNA-mRNA interactions for individual samples as well
 as sample population
}
