% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miracle.R
\name{miracle_mirna}
\alias{miracle_mirna}
\title{miracle_mirna is used to get the ranked result of the targets of a given mirna}
\usage{
miracle_mirna(target_name, seqScore, mirtarList, mirExpr, tarExpr,
  exprFilter = 1, ID_pos = c())
}
\arguments{
\item{target_name}{}

\item{seqScore}{}

\item{mirtarList}{}

\item{mirExpr}{}

\item{tarExpr}{}

\item{exprFilter}{}

\item{ID_pos}{}
}
\description{
miracle_mirna is used to get the ranked result of the targets of a given mirna
}
