% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miracle.R
\name{miracle_multi_target}
\alias{miracle_multi_target}
\title{miracle_multi_target is used to get the result of the prediction result of multiple given mrnas}
\usage{
miracle_multi_target(target_name, seqScore, mirtarList, mirExpr, tarExpr,
  exprFilter = 1, ID_pos = c())
}
\arguments{
\item{target_name}{}

\item{seqScore}{}

\item{mirtarList}{}

\item{mirExpr}{}

\item{tarExpr}{}

\item{exprFilter}{}

\item{ID_pos}{}
}
\description{
miracle_multi_target is used to get the result of the prediction result of multiple given mrnas
}
