% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miracle.R
\name{miracle_score}
\alias{miracle_score}
\title{miracle_score is the main function}
\usage{
miracle_score(seqScore, mirtaList, mirExpr, taExpr, topList)
}
\arguments{
\item{seqScore}{sequence-based information obtained from Targetscan}

\item{mirtaList}{relationship of sample names in mirna expression and sample names in mrna expression}

\item{mirExpr}{mirna  expression data}

\item{taExpr}{mrna expression data}

\item{topList}{the number of top rankers that users are interested in}
}
\value{
miracle score
}
\description{
get the detailed information of the top rankers
by miracle score, both population-level and individual-level
}
\examples{
####### run tests  #######
\dontrun{final_output = miracle_score(seqScore, mirtaList, mirExpr, taExpr, topList)}
}
\author{
Ruidong Li and Han Qu
}
