% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miracle.R
\name{miracle}
\alias{miracle}
\title{miRACLe predicts miRNA-mRNA interactions for individual samples as well as
sample population}
\usage{
miracle(seqScore, sampleMatch, mirExpr, tarExpr, samSelect = NULL,
  exprFilter = 1, OutputSelect = TRUE)
}
\arguments{
\item{seqScore}{A Z x 3 data.frame that contains sequence-based interaction
scores for putative miRNA-mRNA pairs. The first column must contain gene
symbols for mRNAs, the second column must contain miRNA ids for miRNAs, the
third column must contain sequence-based score for the miRNA-mRNA pair}

\item{sampleMatch}{A character K x 2 matrix of paired sample identifiers used
in mirExpr and tarExpr. The first column must contain the sample identifiers
 used in mirExpr and the second column must contain the sample identifiers
 used in tarExpr. Sample identifiers in the same row should be paired.}

\item{mirExpr}{A numeric M x R matrix as the observed expression of M miRNAs
in R samples. NOTE: colnames is required for mirExpr to map each sample to
the sample identifiers in the first column of sampleMatch. rownames is
required for mirExpr to map each miRNA to the miRNA id in the first column
of sampleMatch.}

\item{tarExpr}{A numeric N x T matrix as the observed expression of N mRNAs
in T samples. NOTE: colnames is required for tarExpr to map each sample to
the sample identifiers in the second column of sampleMatch. rownames is
required for tarExpr to map each mRNA to the gene symbol in the first column
 of sampleMatch.}

\item{samSelect}{Optional, a vector of samples that users select to get the
integrated result over those samples. The elements in this vector should be
matched to the sample identifiers in mirExpr. Default is NULL, which means
that all samples in sampleMatch will be analyzed.}

\item{exprFilter}{a filter to remove the lowly expressed miRNAs and mRNAs.
If provided, exprFiter=k, then both miRNAs and mRNAs that are expressed in
less than (1-k)% of the samples in the provided datasets will be filtered
out. Default is 1.}

\item{OutputSelect}{Logical, select “TRUE” to return the top 10 percent
ranked predictions by scores, and “FALSE” to return the whole prediction
result. Default is TRUE.}
}
\value{
An object list containing the following items:

Ind: inferred individual-level prediction result, a 4*N-column

matrixPop: inferred population-level prediction result, a 4-column matrix
}
\description{
Infer miRNA-mRNA interactions using paired miRNA-mRNA expression profile.
miRACLe is based on a random contact model. It integrates genome-wide
expression profiles and sequence-based interaction scores of miRNAs and mRNAs
to calculate the relative probability of effective contacts. We consider
that the more likely an effective contact takes place between two molecules,
the more likely that a regulatory relationship between them exists.
This function supports prediction of miRNA-mRNA interactions at both
individual and population levels.
}
\section{Detail}{

miRACLe predicts miRNA-mRNA interactions for individual samples as well as
sample population. For individual-level prediction, miRACLe inputs a single
paired miRNA-mRNA expression profile and sequence-based scores for putative
miRNA-mRNA pairs and output a list of miRNA-mRNA interactions ranked by
miRACLe scores. When a sample population is of interest, the individualized
miRACLe scores will be averaged over all samples in the population to
determine a population-level miRACLe score, based on which we output the
ranking list of miRNA-mRNA interactions.

In order to run the current version of miRACLe, the users should provide two
data files that describe the expression levels of each miRNA and mRNA. And
one additional file that defines the correspondence of samples between the
miRNA and mRNA data files. All files are tab-delimited ASCII text files. As
for expression data, both microarray profiling and RNA sequencing data are
accepted. To achieve optimal prediction on the sequencing data, log2
transformed normalized read counts (e.g. RSEM) are recommended as the input
for the program. In order for this function to execute correctly, this
expression matrix should be transformed into a non-negative numeric matrix.

miRACLe provides sequence-based interaction scores for putative miRNA-mRNA
pairs. These scores are originally obtained from TargetScan v7.2
(TargetScan_CWCS_cons and TargetScan_CWCS), DIANA-microT-CDS (DIANA_microT_CDS),
MirTarget v4 (MirTarget4), miRanda-mirSVR (miRanda_mirSVR) and compiled by the
developers to fit the model. Default is TargetScan_CWCS_cons. The other scores
can be downloaded \href{https://figshare.com/s/0b7c68cd5152da27a191}{here}.
User can also provide their own seqScore, as long as the format meets the
requirements.
}

\section{Note}{

Sample identifiers in sampleMatch should be a subset of those in mirExpr and tarExpr.
}

\examples{
data(seqScore)   # load the default sequence-based interaction score
data(Test_data)  # load test datasets
mirExpr <- Test_DLBC_miRNA # miRNA expression
tarExpr <- Test_DLBC_mRNA  # mRNA expression
sampleMatch <- Test_DLBC_sampleMatch # sample matching file
sampleSelect = c("TCGA-FA-A4BB-01A-11R-A31S-13", "TCGA-FA-A4XK-01A-11R-A31S-13", "TCGA-FA-A6HN-01A-11R-A31S-13") # samples selected from the test dataset to analyze
final_output <- miracle(seqScore, sampleMatch, mirExpr, tarExpr, samSelect = sampleSelect, exprFilter = 0.8, OutputSelect = FALSE)
final_output$Ind   # Individual-level result
final_output$Pop   # Population-level result
}
