% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miracle.R
\name{miracle_ind}
\alias{miracle_ind}
\title{miRACLe predicts miRNA-mRNA interactions for individual samples}
\usage{
miracle_ind(seqScore, mirExpr, tarExpr, OutputSelect = TRUE)
}
\arguments{
\item{seqScore}{A Z x 3 data.frame that contains sequence-based interaction
scores for putative miRNA-mRNA pairs. The first column must contain gene
symbols for mRNAs, the second column must contain miRNA ids for miRNAs,
the third column must contain sequence-based score for the miRNA-mRNA pair.}

\item{mirExpr}{A numeric vector as the observed expression of M miRNAs. NOTE:
rownames is required for mirExpr_ind to map each miRNA to the miRNA id in
the second column of seqScore.}

\item{tarExpr}{A numeric vector as the observed expression of N mRNAs. NOTE:
rownames is required for tarExpr_ind to map each mRNA to the gene symbol in
the first column of seqScore.}

\item{OutputSelect}{Logical, select “TRUE” to return the top 10 percent
ranked predictions by scores, and “FALSE” to return the whole prediction
result. Default is TRUE}
}
\value{
An object list containing the following items:

Ind: inferred prediction result, a 4-column matrix
}
\description{
Infer miRNA-mRNA interactions using a single paired miRNA-mRNA expression profile.
miRACLe is based on a random contact model. It integrates genome-wide expression
profile and sequence-based interaction scores of miRNAs and mRNAs to calculate
the relative probability of effective contacts. We consider that the more likely
an effective contact takes place between two molecules, the more likely that a
regulatory relationship between them exists. This function is specifically
designed to predict individualized miRNA-mRNA interactions using a single
paired miRNA-mRNA expression profile.
}
\section{Detail}{

miRACLe predicts miRNA-mRNA interactions for individual samples as well as
sample population. For individual-level prediction, miRACLe inputs a single
paired miRNA-mRNA expression profile and sequence-based scores for putative
miRNA-mRNA pairs and output a list of miRNA-mRNA interactions ranked by
miRACLe scores. When a sample population is of interest, the individualized
miRACLe scores will be averaged over all samples in the population to
determine a population-level miRACLe score, based on which we output the
ranking list of miRNA-mRNA interactions.

In order to run the current version of miRACLe, the users should provide two
data files that describe the expression levels of each miRNA and mRNA. And
one additional file that defines the correspondence of samples between the
miRNA and mRNA data files. All files are tab-delimited ASCII text files. As
for expression data, both microarray profiling and RNA sequencing data are
accepted. To achieve optimal prediction on the sequencing data, log2
transformed normalized read counts (e.g. RSEM) are recommended as the input
for the program. In order for this function to execute correctly, this
expression matrix should be transformed into a non-negative numeric matrix.

miRACLe provides sequence-based interaction scores for putative miRNA-mRNA
pairs. These scores are originally obtained from TargetScan v7.2
(TargetScan_CWCS_cons and TargetScan_CWCS), DIANA-microT-CDS (DIANA_microT_CDS),
MirTarget v4 (MirTarget4), miRanda-mirSVR (miRanda_mirSVR) and compiled by the
developers to fit the model. Default is TargetScan_CWCS_cons. The other scores
can be downloaded \href{https://figshare.com/s/0b7c68cd5152da27a191}{here}.
User can also provide their own seqScore, as long as the format meets the
requirements.
}

\examples{
data(seqScore.Rdata) 	# load the default sequence-based interaction score
data(Test_data.Rdata) 	# load test datasets
mirExpr_ind <- Test_HeLa_miRNA	# miRNA expression
tarExpr_ind <- Test_HeLa_mRNA	# mRNA expression
final_output_ind <- miracle_ind(seqScore, mirExpr_ind, tarExpr_ind, OutputSelect = TRUE)
}
